requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
redis==5.0.1
pandas==2.1.3
numpy==1.26.2
pydantic==2.5.0
httpx==0.25.2
pydantic-settings
pydantic-settings
python-dotenv
python-dotenv